# specs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AI Search Portal HTTP API
  version: 0.3.0
  description: >
    Items CRUD and Metadata catalog APIs. Must stay aligned with @ai-search-portal/contracts (Zod).
servers:
  - url: /
    description: Relative to deployment origin
paths:
...
  /api/metadata:
    get:
      operationId: listMetadata
      summary: List metadata assets
      parameters:
        - name: pack
          in: query
          schema:
            type: string
          description: Context pack id (default enterprise-mau)
        - name: q
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListMetadataResponse"
...
  /api/metadata/{assetId}:
    parameters:
      - name: assetId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getMetadataAsset
      summary: Get metadata asset
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetMetadataAssetResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/metadata/{assetId}/lineage:
    get:
      operationId: getMetadataLineage
      summary: Resolve upstream/downstream lineage for an asset
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: string
        - name: pack
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetadataLineageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    MetadataAssetSummary:
      type: object
      required:
        [id, name, description, assetType, owner, tags, classification, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        assetType:
          type: string
          enum: [Database, Table, API, Dashboard]
        owner:
          type: string
        tags:
          type: array
          items:
            type: string
        classification:
          type: string
          enum: [public, internal, PII, confidential]
        updatedAt:
          type: string
        fqn:
          type: string
        facets:
          $ref: "#/components/schemas/DomainFacets"
    MonetaryAmount:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
        currency:
          type: string
          description: ISO 4217 code, e.g. TWD.
          minLength: 3
          maxLength: 3
        per:
          type: string
          description: "Pricing basis: kg, g, session, half_day, unit."
    RegionRef:
      type: object
      required: [code]
      properties:
        code:
          type: string
          description: Stable region code, e.g. TW-TPE, TW-NORTH.
        label:
          type: string
        level:
          type: string
          enum: [country, region, city, district, market]
    TimeWindow:
      type: object
      required: [start]
      properties:
        start:
          type: string
          description: ISO 8601.
        end:
          type: string
          description: ISO 8601; absent means open-ended.
        timezone:
          type: string
          description: IANA zone, e.g. Asia/Taipei.
        recurrence:
          type: string
          description: "Recurrence hint, e.g. seasonal:JUL-SEP, weekly:SAT."
    DomainFacets:
      type: object
      description: Optional money/time/region facet bundle on metadata assets.
      properties:
        domain:
          type: string
          description: Domain key, e.g. agri, metalcraft.
        pricing:
          $ref: "#/components/schemas/MonetaryAmount"
        region:
          $ref: "#/components/schemas/RegionRef"
        timeWindow:
          $ref: "#/components/schemas/TimeWindow"
        completeness:
          type: string
          enum: [complete, partial, minimal]
          description: Modeled completeness signal for partial-but-reasonable data.
...
    ListMetadataResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MetadataAssetSummary"
        pagination:
          type: object
          required: [page, pageSize, total, totalPages]
          properties:
            page:
              type: integer
            pageSize:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
    MetadataAssetDetail:
      allOf:
        - $ref: "#/components/schemas/MetadataAssetSummary"
        - type: object
          required: [fqn, upstreamIds, downstreamIds]
          properties:
            fqn:
              type: string
            upstreamIds:
              type: array
              items:
                type: string
            downstreamIds:
              type: array
              items:
                type: string
    GetMetadataAssetResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/MetadataAssetDetail"
...
    MetadataLineageNode:
      type: object
      required: [id, label, type]
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [table, column, api, dashboard, database]
    MetadataLineageEdge:
      type: object
      required: [source, target]
      properties:
        source:
          type: string
        target:
          type: string
    MetadataLineageResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [assetId, upstream, downstream, nodes, edges]
          properties:
            assetId:
              type: string
            upstream:
              type: array
              items:
                $ref: "#/components/schemas/MetadataAssetSummary"
            downstream:
              type: array
              items:
                $ref: "#/components/schemas/MetadataAssetSummary"
            nodes:
              type: array
              items:
                $ref: "#/components/schemas/MetadataLineageNode"
            edges:
              type: array
              items:
                $ref: "#/components/schemas/MetadataLineageEdge"
...